for CentOS, do scl without sudo
I ran into a problem following these instructions, and this change got me past it.

The command in question enables access to the C++ compiler (puts it on the PATH or something like that).  If you do it with `sudo`, it leaves you at a root shell prompt.  If you then exit that root prompt, to go back to non-root, then g++ is no longer accessible.  If you want to compile as non-root, then you need to do `scl` as non-root.