Fix Disabling AMD GPUs
The AMD jconf.cpp would only accept an array. The config sample, and the nvidia and cpu config work with the value 'null', as they accept a 'kNullType'. This means at current, AMD GPUs could not be turned off, the config file wouldn't parse on load.

This change makes AMD consistent with the others, and can now be disabled.