add independent dev pool coin description
- allow the dev pool to fork on a different block version than the user descriped coin

All algorithm are centered around the user coin description.
It is allowed to have two two different coin algorithms in the user coin description.
It is only allowed to use algorithms for the dev pool coin description those are used in the user coin description.
There are two ways to define a non forking coin.
  - set both user coin algorithm descriptions to the same algorithm and set version to zero
  - set the first algorithm in the user coin description to something you like to use in the dev pool
    and set the second algorithm to the correct representation of the coin. Set the version to 255.
    This will allow that the dev pool can mine on a different coin algorithm than the not forking user coin.
    Do not use an algorithm with different scratchpad size for the dev pool.