fix invalid shares
With rocm we fighted very long with invalid shares. This is now solved with rocm 1.9 and
this tiny fix.
It is not fully clear where a memory optimization is kicking in and break the kernel `Groestl` if the variables `M` and `H` are not `volatile`.
The performance ill not change with this fix.

The fix is tested with rocm 1.9 with a VEGA64 and a RX570