fix crash with monero and strided_index
Strided index 1 is not allowed for cryptonight_v8 and monero.
In the case the dev pool is set to monero and the user tuned there settings for
an other currency the miner will crash if strided index or memChunk is not
fitting the requirement to mine monero.
This PR detects wrong configurations and will set strided index and memChunk to a valid
value but only for cryptonight_v8. The user pool settings will only be changed if monero or
cryptonight_v8 is selected.